% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{covfx}
\alias{covfx}
\title{Return covariance as a function of distance}
\usage{
covfx(r, covModel)
}
\description{
Return covariance as a function of distance
}
\examples{
covModel <- list(kernel="matern",
                 l = 5,     # correlation length
                 v = 1,     # smoothness
                 h = 2.45   # std. deviation
)
r <- seq(0, 20, by = 0.1)
myCov <- covfx(r = r, covModel = covModel)
plot(r, myCov, type = "l", ylim = c(0, max(myCov)),
     ylab = "covariance", xlab = "distance", xaxs = "i", yaxs = "i")
}
