% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GauPro.R
\docType{package}
\name{GauPro-package}
\alias{GauPro}
\alias{GauPro-package}
\title{GauPro: A package for simulating conditional and unconditional Gaussian
Process.}
\description{
The GauPro package provides R functions for Gaussian process (GP) modelling. 
The core functions are coded in C++ and based on the EIGEN library.
}
\section{Features}{

\itemize{
  \item Conditional GP simulation
  \item Space-time GP
  \item GP with monomial mean functions
  \item GP conditioned to derivative observations
}
}
\references{
Several books!
}

