% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariances.R
\name{kernels}
\alias{kernels}
\alias{kGaussian}
\alias{kLinear}
\alias{kMatern}
\title{#' @export
se <- function(x, y, covModel, d = 0, w = 1){
  warning("Deprecated function! Use 'kGaussian' instead!\n")
  kGaussian(r, covModel, d, w)
}
#' @export
kSe <- function(r, covModel, d = 0, w = 1){
  warning("Deprecated function! Use 'kGaussian' instead!\n")
  kGaussian(r = r, para = covModel, d = d, w = w)
}
Kernels (covariance functions) for Gaussian process}
\usage{
kGaussian(r, para, d = 0, w = 1)

kLinear(x, y, para, d = 0, w = 1)

kMatern(r, para, d = 0, w = 1)
}
\description{
Squared Exponential Covariance Function (or radial basis or Gaussian)
over-smoothness, infinitely differentiable at h=0
}
