% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{se}
\alias{se}
\title{Squared Exponential Covariance Function}
\usage{
se(r, covModel, d = 0, w = 1)
}
\description{
Squared Exponential Covariance Function (or radial basis or Gaussian)
over-smoothness, infinitely differentiable at h=0
}

