% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariances.R
\name{crossDist}
\alias{crossDist}
\title{Cross-distance between two matrix}
\usage{
crossDist(X, Y, M = NULL, use_symmetry = use_symmetry)
}
\arguments{
\item{X}{a matrix or vector}

\item{Y}{a matrix or vector with same number of columns as X}

\item{M}{a positive semidefinite matrix (nrow(M) = ncol(M) = ncol(X))}
}
\description{
Compute the distance between every rows of two matrix. The returned distance
has for dimension: nrow(X) x ncol(Y).
If M is the identity matrix (by default), the distance is isotropic, if not
the distance is anisotropic.
}
