% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GauProMod.R
\docType{package}
\name{GauProMod-package}
\alias{GauProMod}
\alias{GauProMod-package}
\title{GauProMod: A package for simulating conditional and unconditional Gaussian
Process.}
\description{
The GauProMod package provides R functions for Gaussian process (GP) modelling. 
The core functions are coded in C++ and based on the EIGEN library.
}
\section{Features}{

\itemize{
  \item Conditional GP simulation
  \item Space-time GP
  \item GP with monomial mean functions
  \item GP conditioned to derivative observations
}
}

\references{
Several books!
}
\author{
\strong{Maintainer}: Emanuel Huber \email{emanuel.huber@alumni.ethz.ch}

Authors:
\itemize{
  \item Emanuel Huber \email{emanuel.huber@alumni.ethz.ch}
}

}
