% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{kernels}
\alias{kernels}
\alias{kGaussian}
\alias{kLinear}
\alias{kMatern}
\title{Kernels (covariance functions) for Gaussian process}
\usage{
kGaussian(r, para, d = 0, w = 1)

kLinear(x, y, para, d = 0, w = 1)

kMatern(r, para, d = 0, w = 1)
}
\description{
Squared Exponential Covariance Function (or radial basis or Gaussian)
over-smoothness, infinitely differentiable at h=0
}
