% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariances.R
\name{kernels}
\alias{kernels}
\alias{kGaussian}
\alias{kLinear}
\alias{kPolynomial}
\alias{kMatern}
\alias{kTriangular}
\alias{kSpherical}
\alias{kCauchy}
\title{Kernels (covariance functions) for Gaussian process}
\usage{
kGaussian(r, para, d = 0, w = 1, use_symmetry = FALSE)

kLinear(x, y, para, d = 0, w = 1, use_symmetry = FALSE)

kPolynomial(x, y, para, d = 0, w = 1, use_symmetry = FALSE)

kMatern(r, para, d = 0, w = 1, use_symmetry = FALSE)

kTriangular(r, para, d = 0, w = 1, use_symmetry = FALSE)

kSpherical(r, para, d = 0, w = 1, use_symmetry = FALSE)

kCauchy(r, para, d = 0, w = 1, use_symmetry = FALSE)
}
\description{
Squared Exponential Covariance Function (or radial basis or Gaussian)
over-smoothness, infinitely differentiable at h=0
}
